requirements.txt
streamlit
torch
torchvision
opencv-python
Pillow
matplotlib
numpy
flask